modified sdfghjklkjhgfdssdfghjk

+++ signup.cpp
@@ -26,7 +26,7 @@ signup::~signup()
 
 void signup::on_pushButton_clicked()
 {
-     if(dbsignup.open())
+    dbsignup.open();
      {
        QString firstname = ui->lineEdit->text();
        QString midname= ui-> lineEdit_17 ->text();
@@ -37,42 +37,59 @@ void signup::on_pushButton_clicked()
        QString conpass=ui->lineEdit_5->text();
        QString dob=ui->dateEdit->text();
        //QCheckBox tandp=ui->checkBox->setCheckState()
-       if(password!=conpass)
+       if(firstname=="\0" || lastname=="\0" || email=="\0" || username=="\0" || password=="\0")
        {
-           QMessageBox::warning(this,"Error","Password doesn't match. Please try again.");
+             QMessageBox::warning(this,"Error","Please fill all the information.");
        }
        else
-        {
-            QSqlQuery qry;
-          qry.prepare("INSERT INTO signup (Firstname, Midname, Lastname, email, username, password, dob)""VALUES(:Firstname, :Midname, :Lastname, :email, :username, :password, :dob)");
-          qry.bindValue(":Firstname", firstname);
-          qry.bindValue(":Midname", midname);
-          qry.bindValue(":Lastname", lastname);
-          qry.bindValue(":email", email);
-          qry.bindValue(":username", username);
-          qry.bindValue(":password", password);
-          qry.bindValue(":dob", dob);
+       {
+       if(ui->checkBox->isChecked())
+       {
 
-          if(qry.exec())
-          {
 
-               QMessageBox::information(this,"Sign Up","Signed up successfully.");
-               usernames=username;
-               QSqlDatabase::removeDatabase("QMYSQL");
-                hide();
-                 basicinfoobj=new basicinfo(this);
-                 basicinfoobj -> show();
-          }
-          else
-          {
-              QMessageBox::warning(this,"Sign Up","Signed up failed.");
-              qDebug() << qry.lastError().text()<<Qt::endl;
-              QSqlDatabase::removeDatabase("QMYSQL");
-          }
+           if(password!=conpass)
+           {
+               QMessageBox::warning(this,"Error","Password doesn't match. Please try again.");
+           }
+           else
+            {
+                QSqlQuery qry;
+              qry.prepare("INSERT INTO signup (Firstname, Midname, Lastname, email, username, password, dob)""VALUES(:Firstname, :Midname, :Lastname, :email, :username, :password, :dob)");
+              qry.bindValue(":Firstname", firstname);
+              qry.bindValue(":Midname", midname);
+              qry.bindValue(":Lastname", lastname);
+              qry.bindValue(":email", email);
+              qry.bindValue(":username", username);
+              qry.bindValue(":password", password);
+              qry.bindValue(":dob", dob);
+
+              if(qry.exec())
+              {
+
+                   QMessageBox::information(this,"Sign Up","Signed up successfully.");
+                   usernames=username;
+                   QSqlDatabase::removeDatabase("QMYSQL");
+                    hide();
+                     basicinfoobj=new basicinfo(this);
+                     basicinfoobj -> show();
+              }
+              else
+              {
+                  QMessageBox::warning(this,"Sign Up","The username is taken.");
+                  qDebug() << qry.lastError().text()<<Qt::endl;
+                  QSqlDatabase::removeDatabase("QMYSQL");
+              }
+            }
         }
+       else
+       {
+           QMessageBox::information(this,"Terms and Conditions","Please accept the Terms and Conditions.");
+       }
+       }
+         dbsignup.close();
 
-      }
-     dbsignup.close();
+    }
+
